# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Conduit API",
    "description": "Multi-project AI chat platform with RBAC, LDAP integration, and pluggable agent connectors",
    "version": "0.1.0",
    "contact": {
      "name": "Conduit Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://conduit.example.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "auth",
      "description": "Authentication and user management"
    },
    {
      "name": "projects",
      "description": "Project management and configuration"
    },
    {
      "name": "chat",
      "description": "Chat streaming and message handling"
    },
    {
      "name": "conversations",
      "description": "Conversation CRUD operations"
    },
    {
      "name": "agents",
      "description": "Agent configuration and management"
    },
    {
      "name": "rbac",
      "description": "Role-based access control"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": ["auth"],
        "summary": "User login",
        "description": "Authenticate user via LDAP and return JWT token",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "User logout",
        "description": "Logout the current user",
        "operationId": "logout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful logout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get current user",
        "description": "Get current authenticated user info",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/projects": {
      "post": {
        "tags": ["projects"],
        "summary": "Create project",
        "description": "Create a new project for the authenticated user",
        "operationId": "createProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/projects/owned": {
      "get": {
        "tags": ["projects"],
        "summary": "List admin projects",
        "description": "List all projects the user can administer (owner or admin role)",
        "operationId": "listAdminProjects",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_name}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get project",
        "description": "Get project details by name",
        "operationId": "getProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          }
        }
      },
      "put": {
        "tags": ["projects"],
        "summary": "Update project",
        "description": "Update project configuration",
        "operationId": "updateProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully"
          }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Delete project",
        "description": "Delete a project (owner only)",
        "operationId": "deleteProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          },
          "403": {
            "description": "Forbidden - owner role required"
          }
        }
      }
    },
    "/chat": {
      "post": {
        "tags": ["chat"],
        "summary": "Stream chat",
        "description": "Stream chat response using Server-Sent Events (SSE)",
        "operationId": "streamChat",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Project-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project name for context"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - project name required"
          },
          "404": {
            "description": "No agent configured"
          }
        }
      }
    },
    "/conversations": {
      "get": {
        "tags": ["conversations"],
        "summary": "List conversations",
        "description": "List all conversations for a project",
        "operationId": "listConversations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Project-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of conversations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Conversation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["conversations"],
        "summary": "Create conversation",
        "description": "Create a new conversation",
        "operationId": "createConversation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Project-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}": {
      "get": {
        "tags": ["conversations"],
        "summary": "Get conversation",
        "description": "Get conversation details and messages",
        "operationId": "getConversation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Project-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["conversations"],
        "summary": "Delete conversation",
        "description": "Delete a conversation and all its messages",
        "operationId": "deleteConversation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Project-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation deleted"
          }
        }
      }
    },
    "/agents": {
      "get": {
        "tags": ["agents"],
        "summary": "List agents",
        "description": "List all agents for a project",
        "operationId": "listAgents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["agents"],
        "summary": "Create agent",
        "description": "Create a new agent configuration",
        "operationId": "createAgent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "john.doe"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "example": "bearer"
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "john.doe"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "john.doe"
          }
        }
      },
      "ProjectCreate": {
        "type": "object",
        "required": ["project_name"],
        "properties": {
          "project_name": {
            "type": "string",
            "example": "my-ai-project"
          },
          "disable_authentication": {
            "type": "boolean",
            "default": false
          },
          "disable_message_storage": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ProjectUpdate": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "disable_authentication": {
            "type": "boolean"
          },
          "disable_message_storage": {
            "type": "boolean"
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "my-ai-project"
          },
          "description": {
            "type": "string"
          },
          "owner_id": {
            "type": "integer"
          },
          "disable_authentication": {
            "type": "boolean"
          },
          "disable_message_storage": {
            "type": "boolean"
          },
          "is_admin": {
            "type": "boolean",
            "description": "Whether current user has admin access"
          }
        }
      },
      "ChatRequest": {
        "type": "object",
        "required": ["conversation_id", "message"],
        "properties": {
          "conversation_id": {
            "type": "string",
            "example": "conv-123"
          },
          "message": {
            "type": "string",
            "example": "What is the weather today?"
          },
          "model": {
            "type": "string",
            "example": "gpt-4"
          },
          "agent_id": {
            "type": "integer",
            "example": 1
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "conv-123"
          },
          "title": {
            "type": "string",
            "example": "Weather Discussion"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ConversationCreate": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "New Conversation"
          }
        }
      },
      "ConversationDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "role": {
            "type": "string",
            "enum": ["user", "assistant", "system"]
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "token_count": {
            "type": "integer"
          }
        }
      },
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "GPT-4 Agent"
          },
          "type": {
            "type": "string",
            "enum": ["langgraph", "openai", "http"],
            "example": "openai"
          },
          "url": {
            "type": "string",
            "example": "https://api.openai.com/v1"
          },
          "is_default": {
            "type": "boolean"
          },
          "project_id": {
            "type": "integer"
          }
        }
      },
      "AgentCreate": {
        "type": "object",
        "required": ["name", "type", "url", "project_id"],
        "properties": {
          "name": {
            "type": "string",
            "example": "My Custom Agent"
          },
          "type": {
            "type": "string",
            "enum": ["langgraph", "openai", "http"]
          },
          "url": {
            "type": "string",
            "example": "https://api.example.com"
          },
          "project_id": {
            "type": "integer"
          },
          "is_default": {
            "type": "boolean",
            "default": false
          },
          "auth": {
            "type": "object",
            "description": "Authentication configuration (auth_type, credentials)"
          },
          "extras": {
            "type": "object",
            "description": "Connector-specific configuration"
          }
        }
      }
    }
  }
}